Added update and delete author.

+++ src/main/java/blue/bookapp/services/AuthorServiceImpl.java
@@ -2,6 +2,7 @@ package blue.bookapp.services;
 
 import blue.bookapp.commands.AuthorCommand;
 import blue.bookapp.converters.AuthorCommandToAuthor;
+import blue.bookapp.converters.AuthorToAuthorCommand;
 import blue.bookapp.domain.Author;
 import blue.bookapp.repositories.AuthorRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -17,10 +18,14 @@ public class AuthorServiceImpl implements AuthorService {
 
     private AuthorRepository authorRepository;
     private AuthorCommandToAuthor authorCommandToAuthor;
+    private AuthorToAuthorCommand authorToAuthorCommand;
+    private BookService bookService;
 
-    public AuthorServiceImpl(AuthorRepository authorRepository, AuthorCommandToAuthor authorCommandToAuthor) {
+    public AuthorServiceImpl(AuthorRepository authorRepository, AuthorCommandToAuthor authorCommandToAuthor, AuthorToAuthorCommand authorToAuthorCommand, BookService bookService) {
         this.authorRepository = authorRepository;
         this.authorCommandToAuthor = authorCommandToAuthor;
+        this.authorToAuthorCommand = authorToAuthorCommand;
+        this.bookService = bookService;
     }
 
     @Override
@@ -41,20 +46,27 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
-    public void removeAuthorById(Long id) {
+    public void removeById(Long id) {
+        Optional<Author> authorOptional = authorRepository.findById(id);
+        if (!authorOptional.isPresent())
+        {
+            throw new RuntimeException("Author not found!");
+        }
+        Author author = authorOptional.get();
+        author.getBooks().forEach(book -> book.setAuthor(null));
+        author.setBooks(null);
         authorRepository.deleteById(id);
     }
 
     @Override
-    public Author findAuthorById(Long id) {
+    public AuthorCommand findCommandById(Long id) {
         Optional<Author> authorOptional = authorRepository.findById(id);
         if (!authorOptional.isPresent())
         {
             throw new RuntimeException("Author with ID not found!");
         }
-        Author author = authorOptional.get();
 
-        return author;
+        return authorToAuthorCommand.convert(authorOptional.get());
     }
 
     @Override
@@ -62,5 +74,18 @@ public class AuthorServiceImpl implements AuthorService {
         Set<Author> authors = new HashSet<>();
         authorRepository.findAll().forEach(authors::add);
         return authors;
+    }
+
+    @Override
+    public AuthorCommand updateAuthor(AuthorCommand authorCommand) {
+        Optional<Author> authorOptional = authorRepository.findById(authorCommand.getId());
+        if (!authorOptional.isPresent())
+        {
+            log.error("Author not found!");
+            throw new RuntimeException("Author not found!");
+        }
+        authorRepository.save(authorCommandToAuthor.convert(authorCommand));
+
+        return authorCommand;
     }
 }
